data/esphome_mqtt_to_influx2/requirements.txt
influxdb-client
paho-mqtt


